Add an echo of the characters sent through the serial port.  It's not perfect, but it's a start.
Also, do some fiddling with the paddle logic.  This one has a natural
mechanism for swapping the paddles for normal/reverse and also makes
it easier to send the KN prosign and still has little difficulty with
E's and the AR prosign.  I'm not sure I can make it any better, so I
guess I'll have to train my fingers.