Playback of memories works, sort of.
The memories are pre-programmed with the memory number, and pushing the
button causes that memory to be played.  There are two issues with the
playback, though.  First, it says the source is "Memory ##" instead of
"Memory 1" or "Memory 2" or whatever.  Second, the characters are not
put into the display as they are sent.