Fix recording to the memory from the keyboard.
The puzzling thing was that sending morse from the keyboard worked just
fine, even when recording the memory, but playing back was broken.
Logging what got put into the memory showed that the issue was with the
recording.

Eventually, figured it out.  The text-to-morse functionality was
handling word spaces by doing key-up twice in a row.  This worked in the
absense of a memory recording because going key up when the system was
already key up didn't result in anything getting keyed.  The problem is
that the memory recording worked on transitions and assumed that new key
up meant previously key down.  So, the end of every word got two key
downs in a row.  Not what I wanted.

I tried fixing up the memory logic to base what it does on the previous
key-up or key-down action rather than the current one.  That sent the
correct morse, but since the key-up actions were not coalesced, it didn't
put a space in the scrolling display even though the recorded message was
sent properly.

So, out of alternatives, I changed the text-to-morse logic to not do a
word space by having an extra transition.  I added a boolean member
variable to the text_to_morse class to indicate whether or not the
system was in a word space and added logic such that the state of the
transmitter wasn't changed until it was out of the word space.  This
made the system behave in the way I wanted.  It could record from the
keyboard (and, I presume, the serial input although I haven't tested it)
and play back what was recorded, and the display included word spaces in
the appropriate places.